Changed Tag Handling
So, now most things rely on a file called tagindex.
It's generated from .augmented files and is rebuilt if any of the mime
files change, or if a new one is added or one is removed.

It generates a file that looks like:
tag1 file1
tag2 file1
tag1 file2
tag3 file2
tag4 file2
tag2 file3

Or whatever.
So, then all.do just uses that index to figure out it should build:
file1.tag1.html file1.tag2.html file2.tag1.html ...

That makes me much happier, because it cut down on the amount of
knowledge all.do needed.
It used to search through mime files to Tag: and stuff.
Now it just generates the tagindex and says "Ok, I'll build these"

The per-tag tagindex is removed in this iteration.
It's now become (grep '^tag ' < tagindex | cut -d ' ' -f2-)
I could have made a tagindex for each tag that just did that, but
I decided that was overkill.

So, default.html.do was changed to use that rather than $tag.tagindex
It got a little uglier... because I end up doing the above 3 times.
I might "cache" it into a variable later.

Deploy didn't change a lot.
I just realized, as I was looking into why it needed to know which
tags a file had, that I wrote it badly.
I was grepping through mime files to try and figure out some list
of all tags, so I could make those directories.
Then I proceded to iterate over all html files and extracted the tag
from the filename as God (I) intended.
So I just took out the part above and put a (mkdir -p) in the iteration
part of the loop, right before the part where I use that directory.
So, now it doesn't require knowledge of tags at all, it just puts
files of the format title.tag.html into tag/title.html.

This makes me much happier.

So, I think this was a good step.