Added More Mime Files
More useless tests.
I built a few of these up as I made sure things like tagindex and
all and stuff were behaving properly.
Things like rebuilding the index if the tags a file's in changes,
or rebuilding the index if I add a new file, but not rebuilding
if I don't do any of those things.

It also gives more things to test next and prev with, and anything
more stressful that 2 files is a better test.