Added Type Conversion
So, now it extracts the Content-Type and builds a type-converter for
that type.
It defaults to text/plain.

I've written a text/plain, which just puts <pre> before the body and
</pre> after the body.

I've also added a text/html which does nothing.
Feasibly it could do cleanup, or sanitizing, or anything, really.

As a test for that one, I changed t4.mime to be html.

So. Yeah.