Removed ID from default.augmented.do
I realized that the ID is the filename I chose.
The only reason I had ID in GitSite was because the filenames were
the title.
Here the title is in the file, and the filename is the unique ID.

So, if I decide I want an ID separate from this later I can add it
back in.
If not, then I'm glad it's gone.