Added Git Augmentation
So, now it uses git dates.
This introduces the possibility for two articles to have exactly the
same creation date.

So, I've changed tagindex to sort first on creation date, then on
filename.

That way we end up with a consistent and reliable ordering when two or
more mime files are added in the same commit.